requirements.txt
langchain
langchain-openai
azure-search-documents
python-dotenv
openai
azure-identity
streamlit
langchain-community
langchain-core
PyYAML